feat: update dto type for match with entity

+++ src/main/java/com/openclassrooms/rental/repository/MessageRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.openclassrooms.rental.model.MessageEntity;
 
-public interface MessageRepository extends JpaRepository<MessageEntity, Long> {}
+public interface MessageRepository extends JpaRepository<MessageEntity, Integer> {}
